Move count_values from function to aggregation
